Added check_FASTA_format and FASTA_cleaning
check_FASTA_format
This function check if a list of FASTA rows contains undesired characters.
If yes, then returns the list of undesired characters found.
Moreover, check if the sequences in the FASTA are represented in singe lines (NO MULTILINES SEQUENCES).

FASTA_cleaning
This function cleans the FASTA file from undesired characters.
In particular remove the '\n' character from the FASTA sequences. This is necessary to further sequencing process.