Modified gen_protein_CDS_index - create index using protein IDs instead gene IDs
The index that link the protein to their CDS from the reference genome annotation file, it is now parsed extracting the 'protein_id' field instead of the 'gene'.
However, for the viral reference genome would be better to use the gene codes and avoid the protein codes conversion when is possible.