Added UniProt to Ensamble conversion management
The software collect the UniProt accession codes and convert in Ensembl codes.

If this conversion table does not exist at all the new one is created.
If the table exists the software compare the UniProt codes already converted, comparing them with the codes provided in input. If some of them are missed in the conversion table, then it performs the conversion and updates the original conversion table with the new entries.