Modified gen_protein_CDS_index - now the index is built based on the annotation format
If the annotations are in GTF format the file is upload in a dataframe.
If the annotations format do NOT COMPLY with the GTF format then the annotations are upload in a numpy matrix.

The RGB tuples report rounded values for each type of color (no more long real numbers)