Modified make_sep_file
The new method to filter the peptide is to maintain the np.array with all the PoGo peptides.
For thos that do not map in the allowed genomic space, the first field of the bed track (chromosome name) is put to ''.
The function will skip the row of the element that report the escape signal in the first position of the element row.