Added add_pep_rgb_inten and fixed the issue of peptides mapped on proteins they do not belong to.
add_pep_rgb_inten now colours the peptides based on their intensity provided by input.

Modified function filter_peptides:
Added two new input variables, peptides_input_table and ptm

peptides_input_table -  allows to filter the peptides and avoid tha a peptide maps on proteins that it does not belong to. Before with the PoGo mapping coordinates and the comparison with protein's CDS there was the issue to confirm peptides position that were not part of the allowed genomic space.

ptm - avoids to overwrite the ptm color code. In IGV the PTMs are visualised as a normal peptide in the peptide track and as ptm coloured by ptm type in the ptm track.