Added dict_numlst_2_dict_strlst and Modified check_input_json
dict_numlst_2_dict_strlst
converts numbers in digits in the kv dictionary where the values (v) are the lists of numbers to convert.

Modified check_input_json
quantitation_method and coloring_method added to the json paramenters
The old intensity values are treated as abundance values and converted to float instead to integers.